fixed textboxes appearing in game
the problem was they were added as actors to the stage in menurenderer,
but they were never removed from the stage. I added a boolean
shouldClear that is set to true each time the menu changes. In the
menurenderer, if the shouldClear is true it will clear the stage of all
actors and then proceed to draw the new menu with only the actors
present on that screen.