More menus
I did all of the old ones but Sign up. They are inconsistent right now -
I started by baking the text for the menus and buttons into the images,
but then I figured I should just do it with text, so the later ones are
all text and the earlier ones are all baked. I will redo the baked ones
when I get back